DEMOS-1538: remove duplicate details page.
I'm not sure how i wound up with two.
