+++ client/src/mock-data/deliverableMocks.ts
@@ -4,7 +4,7 @@ import { DELIVERABLE_DETAIL_HEADER_QUERY } from "pages/deliverables/DeliverableD
 import {
   DELIVERABLE_DETAILS_QUERY,
   DeliverableDetailsManagementDeliverable,
-} from "pages/deliverables/DeliverableDetailsManagementPage";
+} from "pages/DemonstrationDetail/deliverables/DemonstrationDeliverableDetailView";
 import {
   DELIVERABLES_PAGE_QUERY,
   DeliverableTableRow,
